backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
opencv-python>=4.8.0
numpy>=1.26.0,<2.0.0
svgpathtools==1.6.1
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
